회원 관리 구현

+++ src/main/java/kr/kh/team3/app/controller/MemberDeleteServlet.java
@@ -14,6 +14,7 @@ import org.json.JSONObject;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import kr.kh.team3.app.model.vo.CategoryVO;
+import kr.kh.team3.app.model.vo.MemberVO;
 import kr.kh.team3.app.pagination.Criteria;
 import kr.kh.team3.app.pagination.PageMaker;
 import kr.kh.team3.app.service.CategoryService;
@@ -26,12 +27,34 @@ public class MemberDeleteServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 	private MemberService memberService = new MemberServiceImp();
 
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		//삭제할 회원 아이디를 가져옴
-		String me_id = request.getParameter("me_id");
+		String me_id = request.getParameter("id");
 		System.out.println(me_id);
-		boolean res = memberService.deleteMember(me_id);
-		response.getWriter().write(res ? "ok" : "");
+		
+		boolean res;
+		//관리자 자신인지 확인
+		MemberVO user = (MemberVO) request.getSession().getAttribute("user");
+		if(user.getMe_id().equals(me_id)) {
+			res = false;
+			System.out.println(user.getMe_id());
+		}else {
+			res = memberService.deleteMember(me_id);
+		}
+		
+		
+		//삭제했으면 삭제했다고 알리고 회원 관리로 이동
+		if(res) {
+			request.setAttribute("msg", "삭제되었습니다.");
+			request.setAttribute("url", "admin/membermanager");
+		}
+		//실패했으면 실패했다고 알리고 게시글 상세로 이동
+		else {
+			request.setAttribute("msg", "삭제를 실패했습니다.");
+			request.setAttribute("url", "admin/membermanager");
+		}
+		
+		request.getRequestDispatcher("/WEB-INF/view/message.jsp").forward(request, response);
 	}
 
 }
